pyproject.toml
[tool.poetry]
name = "fastapi-app"
version = "0.1.0"
description = ""
authors = ["Sergey <mudgtt@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
uvicorn = {extras = ["standard"], version = "^0.25.0"}
fastapi = "^0.109.0"
pydantic-settings = "^2.1.0"
alembic = "^1.13.1"
psycopg2-binary = "^2.9.9"
sqladmin = {extras = ["full"], version = "^0.16.0"}
asyncpg = "^0.29.0"
sqlmodel = "^0.0.14"
gunicorn = "^21.2.0"
fastapi-users = {extras = ["sqlalchemy"], version = "^12.1.2"}
pyjwt = {extras = ["crypto"], version = "^2.8.0"}
bcrypt = "^4.1.2"
multipart = "^0.2.4"


[tool.poetry.group.dev.dependencies]
pytest = "^7.4.4"
black = "^23.12.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
